app/airm_module/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
python-multipart==0.0.20
pdfplumber==0.11.7
pdfminer.six==20250506
python-docx==1.2.0
lxml==6.0.2
pypdfium2==4.30.0
